Moved Mutex.cs to shared
